Actions and reducers were split in different folders
