+++ src/reducers/reducer.test.jsx
@@ -1,14 +1,6 @@
-import {CHANGE_FILTER, reducer, initialState, ActionCreator} from './reducer';
-import {FilterNames} from './consts/consts';
-
-describe(`Action creators work correctly`, () => {
-  it(`Action creator for changing filter returns correct action`, () =>
-    expect(ActionCreator.changeFilter(FilterNames.KIDS_AND_FAMILY)).toEqual({
-      type: CHANGE_FILTER,
-      payload: FilterNames.KIDS_AND_FAMILY
-    })
-  );
-});
+import {reducer, initialState} from './reducer';
+import {CHANGE_FILTER} from '../actions/actions';
+import {FilterNames} from '../consts/consts';
 
 describe(`Reducer works correctly`, () => {
   it(`Default reducer returned state`, () =>
